+++ app/(url)/page.tsx
@@ -1,4 +1,9 @@
 import ConvertUrlForm from "@/components/convert-url-form";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "홈",
+};
 
 export default async function Home() {
   return (
